requirements.txt
astroid==2.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.2
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
pyusb==1.0.2
rope==0.17.0
shiboken6==6.8.0.1
six==1.14.0
toml==0.10.0
wrapt==1.12.1
